pyproject.toml
[tool.poetry]
name = "miniopenaitelegrambot"
version = "0.1.0"
description = ""
authors = ["LagrangeH <maxsergeev39@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.6.0"
python-telegram-bot = "^20.0"
environs = "^9.5.0"
openai = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
